Hide dependency on the json library by using a factory method
