+++ src/WinTrek/WinTrek.cpp
@@ -719,7 +719,7 @@ public:
 TRef<TrekWindow> g_ptrekWindow;
 TRef<UpdatingConfiguration> g_pConfiguration = new UpdatingConfiguration(
     std::make_shared<FallbackConfigurationStore>(
-        std::make_shared<JsonConfigurationStore>(GetExecutablePath() + "\\config.json"),
+        CreateJsonConfigurationStore(GetExecutablePath() + "\\config.json"),
         std::make_shared<RegistryConfigurationStore>(HKEY_CURRENT_USER, ALLEGIANCE_REGISTRY_KEY_ROOT)
         )
 );
